moving witness out of interfaces
