+++ yarn-project/circuit-types/src/index.ts
@@ -6,6 +6,7 @@ export * from './l2_block.js';
 export * from './body.js';
 export * from './l2_block_downloader/index.js';
 export * from './l2_block_source.js';
+export * from './public_data_witness.js';
 export * from './tx_effect.js';
 export * from './logs/index.js';
 export * from './merkle_tree_id.js';
