+++ yarn-project/circuit-types/src/interfaces/aztec-node.ts
@@ -20,6 +20,7 @@ import {
   type LogType,
 } from '../logs/index.js';
 import { type MerkleTreeId } from '../merkle_tree_id.js';
+import { type PublicDataWitness } from '../public_data_witness.js';
 import { type SiblingPath } from '../sibling_path/index.js';
 import { type ProcessOutput, type Tx, type TxHash, type TxReceipt } from '../tx/index.js';
 import { type TxEffect } from '../tx_effect.js';
@@ -27,7 +28,6 @@ import { type SequencerConfig } from './configs.js';
 import { type L2BlockNumber } from './l2_block_number.js';
 import { type NullifierMembershipWitness } from './nullifier_tree.js';
 import { type ProverConfig } from './prover-client.js';
-import { type PublicDataWitness } from './public_data_witness.js';
 
 /**
  * The aztec node.
