+++ yarn-project/circuit-types/src/interfaces/index.ts
@@ -4,7 +4,6 @@ export * from './pxe.js';
 export * from './sync-status.js';
 export * from './configs.js';
 export * from './nullifier_tree.js';
-export * from './public_data_witness.js';
 export * from './prover-client.js';
 export * from './proving-job.js';
 export * from './block-prover.js';
